+++ srcs/commands/ft_echo.c
@@ -1,0 +1,39 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_echo.c                                          :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: vmoreau <vmoreau@student.42.fr>            +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2020/10/22 16:16:29 by vmoreau           #+#    #+#             */
+/*   Updated: 2020/10/22 16:17:55 by vmoreau          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../includes/minishell.h"
+
+char	*ft_echo(t_list_cmd *args)
+{
+	int		nflag;
+	char	*ret;
+
+	nflag = 0;
+	if (!args || !args->str)
+		return (ft_strdup("\n"));
+	while (!ft_strcmp(args->str, "-n"))
+	{
+		nflag = 1;
+		args = args->next;
+	}
+	ret = ft_strdup("");
+	while (args)
+	{
+		ret = ft_strjoin_free(ret, args->str);
+		if (args->next && !(args->flags & F_NO_SPACE_AFTER))
+			ret = ft_strjoin_free(ret, " ");
+		args = args->next;
+	}
+	if (!nflag)
+		ret = ft_strjoin_free(ret, "\n");
+	return (ret);
+}
