+++ srcs/commands/ft_env.c
@@ -1,0 +1,28 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_env.c                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: vmoreau <vmoreau@student.42.fr>            +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2020/10/22 16:16:31 by vmoreau           #+#    #+#             */
+/*   Updated: 2020/10/22 16:19:15 by vmoreau          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../includes/minishell.h"
+
+char	*ft_env(char **envp)
+{
+	char	*ret;
+	int		i;
+
+	ret = ft_strdup("");
+	i = 0;
+	while (envp[i])
+	{
+		ret = ft_strjoin_free(ret, envp[i++]);
+		ret = ft_strjoin_free(ret, "\n");
+	}
+	return (ret);
+}
