+++ srcs/commands/ft_exit.c
@@ -1,0 +1,26 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_exit.c                                          :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: vmoreau <vmoreau@student.42.fr>            +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2020/10/22 16:16:34 by vmoreau           #+#    #+#             */
+/*   Updated: 2020/10/22 16:18:34 by vmoreau          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../includes/minishell.h"
+
+char	*ft_exit(t_list_cmd *args)
+{
+	int	err;
+
+	if (args)
+	{
+		if ((err = ft_atoi(args->str)))
+			exit(err);
+	}
+	exit(0);
+	return (NULL);
+}
